+++ packages/ubq/CHANGELOG.md
@@ -6,8 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+ - `@unitybase/ubq/modules/mail-queue` exports a property `mailerEnabled`.
+   - Indicate mailer is configured in `serverConfig.application.customSettings.mailerConfig`.
+   - in case this property is false calls to `queueMail` do nothing, so better to verify it before mail creation to save a server resources
 
 ### Changed
+ - `queueMail` do not put a mail sending job into queue in case mailer is not configured in `serverConfig.application.customSettings.mailerConfig` 
 
 ### Deprecated